Modified the data record ComponentTreeNode in clj-ctree.core, with changes to clj-ctree.image, clj-ctree.utils, and clj-ctree.vectors.
Added a bounding box to ComponentTreeNode (as field :box), which is implemented as a vector to two raster offsets, and changed the value of :variance from a scalar to a matrix of second-moments represented as a flat vector of length n(n+1)/2.  Also, the field :mean was changed semantically, to represent the mean lattice site of the component instead of its spatial position. Substantially updated cjl-ctree.core/merge-ctrees, and replaced ctree-contains-offset? with ctree-contains-offset-at-site? to exploit the existence of bounding boxes. Also changed clj-ctree.core/make-summary to estimate the covariance matrix of the lattice sites in each component.

In clj-ctree.image, replaced the ImageAccess function get-position-of-site with scale-site, though its function is unchanged, and created a new function get-scale-factors, which returns the physical dimensions of each voxel.

Created clj-ctree.utils/square for the sake of efficiency, perhaps.

created clj-ctree.vectors/flat-vector-auto-product which is used to compute the second-order correlations from a vector.